Update README.md
Agriculture being the backbone of the Indian Economy faces multiple challenges in the food production and prediction sectors due to the unnatural climatic changes, which reduces crop yield adversely. Farmers lack the knowledge and help to forecast the future crop yield and take the necessary precautions. This project aims to guide farmers in the choice of crops to be grown by predicting the suitable crops for a given region by incorporating supervised machine learning algorithms which are one of the advanced technologies in crop prediction.